BackendArchiecture/requirements.txt
google-generativeai>=0.3.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0
pathlib>=1.0.1
streamlit>=1.32.0 